requirements-arm64.txt
# For Oracle Database support:
# see https://oracle.github.io/odpi/doc/installation.html#linux

arrow==0.15.7
celery[redis]==4.4.6
clickhouse-driver==0.1.3
croniter==1.0.6
cx-Oracle==7.3.0
DBUtils==1.3
funcsigs==1.0.2
influxdb==5.2.3
jsonpath2==0.4.5
Markdown==3.3.6
nanoid==2.0.0
paho-mqtt==1.5.1
psutil==5.8.0
psycopg2==2.8.5
pycryptodomex==3.8.0
pymongo==3.10.1
# pymssql==2.2.1
PyMySQL[rsa,ed25519]==1.0.2
python-dateutil==2.7.2
python3-memcached==1.51
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
retry==0.9.2
simplejson==3.17.0
six==1.12.0
ujson==4.2.0
xmltodict==0.9.2
